+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/elements/BoostedMark.kt
@@ -22,6 +22,7 @@ package com.vitorpamplona.amethyst.ui.note.elements
 
 import androidx.compose.runtime.Composable
 import com.vitorpamplona.amethyst.R
+import com.vitorpamplona.amethyst.commons.ui.note.QuietMark
 import com.vitorpamplona.amethyst.ui.stringRes
 
 @Composable
